requirements.txt
Flask
pandas
qrcode
pillow
gunicorn
psycopg2-binary
